+++ ft_strlcat.c
@@ -6,17 +6,17 @@
 /*   By: lraffin <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/04/11 00:29:40 by lraffin           #+#    #+#             */
-/*   Updated: 2021/04/17 23:50:05 by lraffin          ###   ########.fr       */
+/*   Updated: 2021/11/11 17:51:24 by lraffin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t	ft_strlcat(char *dst, const char *src, size_t dstsize)
+int	ft_strlcat(char *dst, const char *src, int dstsize)
 {
-	size_t	d_len;
-	size_t	s_len;
-	size_t	i;
+	int	d_len;
+	int	s_len;
+	int	i;
 
 	d_len = ft_strlen(dst);
 	s_len = ft_strlen(src);
